requirements.txt
PyMySQL==0.10.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
python-dotenv==0.15.0
